stream-output/server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Vasek Mlejnsky <vasek.mlejnsky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
e2b = "^0.13.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
websockets = "^12.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "server.main:main"
